pyproject.toml
[project]
name = "tradingpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "debugpy>=1.8.20",
  "ipykernel>=7.2.0",
  "jupyter-client>=8.8.0",
  "metatrader5>=5.0.5640",
  "pandas>=3.0.1",
  "polars-lts-cpu>=1.33.1",
  "pydantic>=2.12.5",
  "pynvim>=0.6.0",
  "python-dotenv>=1.2.2",
  "python-telegram-bot>=22.6",
]
